+++ lib/resty/session/strategies/regenerate.lua
@@ -25,8 +25,8 @@ function strategy.touch(session, close)
 end
 
 function strategy.save(session, close)
-  local storage = session.storage
   if session.present then
+    local storage = session.storage
     if storage.ttl then
       storage:ttl(session.encoder.encode(session.id), session.cookie.discard, true)
     elseif storage.close then