+++ lib/resty/session.lua
@@ -684,7 +684,9 @@ end
 
 function session:regenerate(flush, close)
     close = close ~= false
-    regenerate(self, flush)
+    if not self.strategy.regenerate then
+        regenerate(self, flush)
+    end
     return save(self, close)
 end